# Avoid unnecessary checks when calling readCodePointValue on Buffer